+++ runtime/interoperability/__init__.py
@@ -1,3 +1,6 @@
+from .identity import (AuthorityPolicy, ConflictStatus, ConflictValue,
+                       EntityResolution, KnowledgeConflict,
+                       ReconciliationState, ResolutionCandidate, TieBreaker)
 from .model import (AdapterCapability, EntityLocality, Exchange, Extension,
                     ExternalIdentifier, ExternalSystem, GovernanceStatus,
                     ImportMode, IntegrationAdapter, IntegrationErrorCode,
@@ -11,4 +14,7 @@ __all__ = [
     "IntegrationAdapter", "IntegrationErrorCode", "InteropError", "Lossiness",
     "MappingConfidence", "MappingRelation", "Namespace", "SemanticMapping",
     "SyncDirection", "split_concept_ref", "InteropRegistry",
+    "AuthorityPolicy", "ConflictStatus", "ConflictValue", "EntityResolution",
+    "KnowledgeConflict", "ReconciliationState", "ResolutionCandidate",
+    "TieBreaker",
 ]